+++ src/ui/panes/queue.rs
@@ -32,10 +32,11 @@ use crate::{
 };
 use log::error;
 use ratatui::{
+    layout::Flex,
     prelude::{Constraint, Layout, Rect},
     style::Stylize,
     text::Line,
-    widgets::{Block, Borders, Padding, Row, Table, TableState},
+    widgets::{Block, Borders, Row, Table, TableState},
     Frame,
 };
 
@@ -80,16 +81,49 @@ impl Pane for QueuePane {
         let AppContext { queue, config, .. } = context;
         let queue_len = queue.len();
 
-        let header_height = u16::from(config.theme.show_song_table_header);
-        let [table_header_section, mut queue_section] =
-            *Layout::vertical([Constraint::Min(header_height), Constraint::Percentage(100)]).split(area)
-        else {
-            return Ok(());
+        let title = self
+            .filter
+            .as_ref()
+            .map(|v| format!("[FILTER]: {v}{} ", if self.filter_input_mode { "█" } else { "" }));
+
+        let table_block = {
+            let mut b = Block::default().border_style(config.as_border_style().bold());
+            if config.theme.show_song_table_header {
+                b = b.borders(Borders::TOP);
+            }
+            if let Some(ref title) = title {
+                b = b.title(title.clone().blue());
+            }
+            b
         };
+
+        let header_height = u16::from(config.theme.show_song_table_header);
+
+        let [data_area, scrollbar_area] =
+            Layout::horizontal([Constraint::Percentage(100), Constraint::Length(1)]).areas(area);
+
+        let [table_header_section, mut queue_section] =
+            Layout::vertical([Constraint::Length(header_height), Constraint::Min(0)])
+                .horizontal_margin(1)
+                .areas(data_area);
+
+        let constraints: &[Constraint] = if config.theme.show_song_table_header {
+            &[Constraint::Length(header_height + 1), Constraint::Min(0)]
+        } else {
+            &[Constraint::Min(0)]
+        };
+        let scrollbar_index = usize::from(config.theme.show_song_table_header);
+        let scrollbar_area = Layout::vertical(constraints).split(scrollbar_area)[scrollbar_index];
+
+        let table_area = table_block.inner(queue_section);
 
         self.scrolling_state.set_content_len(Some(queue_len));
 
-        let widths = Layout::horizontal(&self.column_widths).split(table_header_section);
+        let widths = Layout::horizontal(self.column_widths.clone())
+            .flex(Flex::Start)
+            .spacing(1)
+            .split(table_area);
+
         let formats = &config.theme.song_table_format;
 
         let table_items = queue
@@ -101,7 +135,9 @@ impl Pane for QueuePane {
                     .is_some_and(|v| v == song.id);
 
                 let columns = (0..formats.len()).map(|i| {
-                    song.as_line_ellipsized(formats[i].prop, widths[i].width.into())
+                    let max_len: usize = widths[i].width.into();
+
+                    song.as_line_ellipsized(formats[i].prop, max_len, &config.theme.symbols)
                         .unwrap_or_default()
                         .alignment(formats[i].alignment.into())
                 });
@@ -121,8 +157,6 @@ impl Pane for QueuePane {
             })
             .collect_vec();
 
-        let mut table_padding = Padding::right(2);
-        table_padding.left = 1;
         if config.theme.show_song_table_header {
             let header_table = Table::default()
                 .header(Row::new(self.header.iter().enumerate().map(|(idx, title)| {
@@ -130,31 +164,14 @@ impl Pane for QueuePane {
                 })))
                 .style(config.as_text_style())
                 .widths(self.column_widths.clone())
-                .block(config.as_header_table_block().padding(table_padding));
+                .block(config.as_header_table_block());
             frame.render_widget(header_table, table_header_section);
         }
 
-        let title = self
-            .filter
-            .as_ref()
-            .map(|v| format!("[FILTER]: {v}{} ", if self.filter_input_mode { "█" } else { "" }));
-        let table_block = {
-            let mut b = Block::default()
-                .padding(table_padding)
-                .border_style(config.as_border_style().bold());
-            if config.theme.show_song_table_header {
-                b = b.borders(Borders::TOP);
-            }
-            if let Some(ref title) = title {
-                b = b.title(title.clone().blue());
-            }
-            b
-        };
         let table = Table::new(table_items, self.column_widths.clone())
             .style(config.as_text_style())
             .row_highlight_style(config.theme.current_item_style);
 
-        let table_area = table_block.inner(queue_section);
         self.table_area = table_area;
         frame.render_stateful_widget(table, table_area, self.scrolling_state.as_render_state_ref());
         frame.render_widget(table_block, queue_section);
@@ -166,7 +183,7 @@ impl Pane for QueuePane {
         self.scrolling_state.set_viewport_len(Some(queue_section.height.into()));
         frame.render_stateful_widget(
             config.as_styled_scrollbar(),
-            queue_section,
+            scrollbar_area,
             self.scrolling_state.as_scrollbar_state_ref(),
         );